Use pipenv for dependency management; bump dependencies to Py37 support
This commit makes a few broad changes to how dependencies are managed.

First, pipenv is now used to manage dependencies in the project. This
simplifies development work. All dependencies are specified in the
Pipfile. For cross-compatibility, an updated requirements.txt file is
distributed alongside too.

Second, I manually reviewed all dependencies on the project. Almost all
of the dependencies were upgraded to include Python 3.7 support. I
bumped the dependency versions since it is a goal to build this project
with Python 3.7 support going forward.

* Add python-fedora dependency